add Swagger and DataBaseClient

+++ src/main/java/br/com/pedrosa/reactive/r2dbc/repository/PersonRepository.java
@@ -3,6 +3,6 @@ package br.com.pedrosa.reactive.r2dbc.repository;
 import br.com.pedrosa.reactive.r2dbc.domain.Person;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 
-public interface PersonRepository extends ReactiveCrudRepository<Person, Long> {
+public interface PersonRepository extends ReactiveCrudRepository<Person, Integer> {
 
 }
